Remove allocations from creating the matrix
Well, it's been a long road, but we finally
have a version of
matrix_of_multiply_then_split_sortmodp
that barely allocates anything.

The function now seems to take about 0.25 seconds
on my device for p=5 test cases,
though it's a bit inconsistent.
This is definitely best so far, but I'm not sure if
it's quite enough to get a surface of height 10.

Right now the code is a bit messy, in particular
I've changed all the slices to for loops.
HybridArrays.jl might be able to make this clearer.